refactor: extract escapeHtml, formatDuration, sanitizeTopicStatus to shared utils/domHelpers.ts
These three utility functions were copy-pasted verbatim between popup.ts
and dashboard.ts. Any bug fix or change (especially to the XSS-critical
escapeHtml) required updating two files. The implementations had already
started to diverge: dashboard.ts escapeHtml only accepted string, popup.ts
accepted string|null|undefined; sanitizeTopicStatus used different syntax.

Extract to src/utils/domHelpers.ts with the best signature from either file,
import in both files, remove the local definitions.

Fixes #492
